requirements.txt
asgiref==3.2.3
certifi==2020.4.5.1
Django==3.0.3
django-cors-headers==3.2.1
django-crispy-forms==1.8.1
djangorestframework==3.11.0
Jinja2==2.11.1
joblib==0.15.1
MarkupSafe==1.1.1
Pillow==7.1.2
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
threadpoolctl==2.1.0
wincertstore==0.2
